+++ src/main/java/fr/adaming/forum/dao/SkillDaoImpl.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 import fr.adaming.forum.entity.Skill;
 
 @Component
-public class SkillDaoImpl implements ISkillDao{
+public class SkillDaoImpl implements ISkillDao {
 
 	@PersistenceContext
 	private EntityManager em;
@@ -31,7 +31,7 @@ public class SkillDaoImpl implements ISkillDao{
 	public List<Skill> getSkillByKeyWord(String keyWord) {
 		Query query = em.createQuery("FROM Skill s where s.skillName like :x OR s.skillLevel like :x");
 		query.setParameter("x", "%" + keyWord + "%");
-		log.info("Il y a " + query.getResultList().size() + " compétences qui inclue(nt) le mot clé " +keyWord );
+		log.info("Il y a " + query.getResultList().size() + " compétences qui inclue(nt) le mot clé " + keyWord);
 		return query.getResultList();
 	}
 
@@ -45,8 +45,10 @@ public class SkillDaoImpl implements ISkillDao{
 	@Override
 	public Skill deleteSkill(Long idSkill) {
 		Skill skill = em.find(Skill.class, idSkill);
-		em.remove(skill);
-		log.info("La compétence " + skill.getSkillName() + " a bien été supprimée!");
+		if (skill != null) {
+			em.remove(skill);
+			log.info("La compétence " + skill.getSkillName() + " a bien été supprimée!");
+		}
 		return skill;
 	}
 
@@ -54,17 +56,17 @@ public class SkillDaoImpl implements ISkillDao{
 	@Override
 	public List<Skill> getAllSkills() {
 		Query query = em.createQuery("FROM Skill");
-		log.info("Il y a "+ query.getResultList().size()+"compétences");
+		log.info("Il y a " + query.getResultList().size() + "compétences");
 		return query.getResultList();
 	}
-	
+
 	@Override
 	public Skill getSkillById(Long idSkill) {
 		Skill skill = em.find(Skill.class, idSkill);
-		if(skill != null){
-			log.info("La compétence "+ skill.getSkillName() +" est dans la base de donnée" );
-		}else{
-			log.info("La compétence "+ idSkill +" n'est pas dans la base de donnée" );
+		if (skill != null) {
+			log.info("La compétence " + skill.getSkillName() + " est dans la base de donnée");
+		} else {
+			log.warning("La compétence " + idSkill + " n'est pas dans la base de donnée");
 		}
 		return skill;
 	}
